requirements.txt
pysam==0.15.2
PyVCF==0.6.8
wheel==0.29.0
